requirements.txt
altgraph==0.17.4
delocate==0.11.0
filelock==3.14.0
fsspec==2024.5.0
Jinja2==3.1.4
macholib==1.16.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pillow==9.5.0
setuptools==70.0.0
sympy==1.12.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
typing_extensions==4.12.0
wheel==0.43.0
